# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeFrameBody {
  /** Base64-encoded image data (JPEG/PNG) from a video frame capture */
  imageBase64: string;
}

/**
 * Severity level of detected threat
 */
export type FrameAnalysisResultThreatLevel =
  (typeof FrameAnalysisResultThreatLevel)[keyof typeof FrameAnalysisResultThreatLevel];

export const FrameAnalysisResultThreatLevel = {
  none: "none",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface FrameAnalysisResult {
  /** Whether fraud indicators were detected */
  isFraudSuspected: boolean;
  /** Severity level of detected threat */
  threatLevel: FrameAnalysisResultThreatLevel;
  /** List of detected fraud indicators (e.g. "Police uniform", "Handcuffs visible") */
  detectedIndicators: string[];
  /** Brief explanation of what was detected */
  explanation: string;
  /** Recommended action for the user */
  recommendation: string;
}

export interface CheckScamNumberBody {
  /** Phone number to check (with or without country code) */
  phoneNumber: string;
}

export interface ScamNumberCheckResult {
  isKnownScam: boolean;
  /** Number of times this number has been reported */
  reportCount: number;
  /** Date when first reported (ISO string) */
  firstReported?: string | null;
  /** Date when last reported (ISO string) */
  lastReported?: string | null;
  /** Type of fraud associated with this number */
  fraudType?: string | null;
  /** Message to show the user */
  warningMessage: string;
}

export interface ReportScamNumberBody {
  phoneNumber: string;
  /** Type of fraud (digital_arrest, impersonation, extortion, etc.) */
  fraudType: string;
  /** Optional description of the incident */
  description?: string | null;
}

export interface ReportScamNumberResult {
  success: boolean;
  message: string;
  reportId: number;
}

export type KnownScamNumbersListNumbersItem = {
  id: number;
  phoneNumber: string;
  reportCount: number;
  fraudType: string;
  lastReported: string;
};

export interface KnownScamNumbersList {
  numbers: KnownScamNumbersListNumbersItem[];
  total: number;
}