# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Uses AI vision to detect police uniforms, police backgrounds, advocate robes, or government office setups in a captured frame
 * @summary Analyze a video frame for fraud indicators
 */
export const AnalyzeFrameBody = zod.object({
  imageBase64: zod
    .string()
    .describe(
      "Base64-encoded image data (JPEG\/PNG) from a video frame capture",
    ),
});

export const AnalyzeFrameResponse = zod.object({
  isFraudSuspected: zod
    .boolean()
    .describe("Whether fraud indicators were detected"),
  threatLevel: zod
    .enum(["none", "low", "medium", "high", "critical"])
    .describe("Severity level of detected threat"),
  detectedIndicators: zod
    .array(zod.string())
    .describe(
      'List of detected fraud indicators (e.g. \"Police uniform\", \"Handcuffs visible\")',
    ),
  explanation: zod.string().describe("Brief explanation of what was detected"),
  recommendation: zod.string().describe("Recommended action for the user"),
});

/**
 * Checks against a database of reported digital arrest scam numbers
 * @summary Check if a phone number is a known scam number
 */
export const CheckScamNumberBody = zod.object({
  phoneNumber: zod
    .string()
    .describe("Phone number to check (with or without country code)"),
});

export const CheckScamNumberResponse = zod.object({
  isKnownScam: zod.boolean(),
  reportCount: zod
    .number()
    .describe("Number of times this number has been reported"),
  firstReported: zod
    .string()
    .nullish()
    .describe("Date when first reported (ISO string)"),
  lastReported: zod
    .string()
    .nullish()
    .describe("Date when last reported (ISO string)"),
  fraudType: zod
    .string()
    .nullish()
    .describe("Type of fraud associated with this number"),
  warningMessage: zod.string().describe("Message to show the user"),
});

/**
 * Report a phone number used in digital arrest fraud
 * @summary Report a new scam number
 */
export const ReportScamNumberBody = zod.object({
  phoneNumber: zod.string(),
  fraudType: zod
    .string()
    .describe("Type of fraud (digital_arrest, impersonation, extortion, etc.)"),
  description: zod
    .string()
    .nullish()
    .describe("Optional description of the incident"),
});

/**
 * Returns paginated list of reported scam numbers
 * @summary Get list of known scam numbers
 */
export const GetKnownScamNumbersResponse = zod.object({
  numbers: zod.array(
    zod.object({
      id: zod.number(),
      phoneNumber: zod.string(),
      reportCount: zod.number(),
      fraudType: zod.string(),
      lastReported: zod.string(),
    }),
  ),
  total: zod.number(),
});