# lib/api-zod/src/generated/types/analyzeFrameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeFrameBody {
  /** Base64-encoded image data (JPEG/PNG) from a video frame capture */
  imageBase64: string;
}