# lib/api-zod/src/generated/types/checkScamNumberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckScamNumberBody {
  /** Phone number to check (with or without country code) */
  phoneNumber: string;
}