# lib/api-zod/src/generated/types/frameAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FrameAnalysisResultThreatLevel } from "./frameAnalysisResultThreatLevel";

export interface FrameAnalysisResult {
  /** Whether fraud indicators were detected */
  isFraudSuspected: boolean;
  /** Severity level of detected threat */
  threatLevel: FrameAnalysisResultThreatLevel;
  /** List of detected fraud indicators (e.g. "Police uniform", "Handcuffs visible") */
  detectedIndicators: string[];
  /** Brief explanation of what was detected */
  explanation: string;
  /** Recommended action for the user */
  recommendation: string;
}