# lib/api-zod/src/generated/types/knownScamNumbersList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KnownScamNumbersListNumbersItem } from "./knownScamNumbersListNumbersItem";

export interface KnownScamNumbersList {
  numbers: KnownScamNumbersListNumbersItem[];
  total: number;
}