# lib/api-zod/src/generated/types/reportScamNumberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportScamNumberBody {
  phoneNumber: string;
  /** Type of fraud (digital_arrest, impersonation, extortion, etc.) */
  fraudType: string;
  /** Optional description of the incident */
  description?: string | null;
}