# lib/api-zod/src/generated/types/scamNumberCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScamNumberCheckResult {
  isKnownScam: boolean;
  /** Number of times this number has been reported */
  reportCount: number;
  /** Date when first reported (ISO string) */
  firstReported?: string | null;
  /** Date when last reported (ISO string) */
  lastReported?: string | null;
  /** Type of fraud associated with this number */
  fraudType?: string | null;
  /** Message to show the user */
  warningMessage: string;
}